MAINT: Reorganize fft/fftpack helper functions
