+++ scipy/fft/_pocketfft/__init__.py
@@ -2,6 +2,7 @@
 
 from .basic import *
 from .realtransforms import *
+from .helper import *
 
 from scipy._lib._testutils import PytestTester
 test = PytestTester(__name__)
