+++ scipy/fft/_realtransforms.py
@@ -1,4 +1,3 @@
-import scipy.fftpack as _fftpack
 from ._basic import _dispatch
 from scipy._lib.uarray import Dispatchable
 import numpy as np
