+++ scipy/fftpack/basic.py
@@ -12,80 +12,6 @@ from numpy import swapaxes, zeros
 from scipy.fft import _pocketfft
 
 
-def _is_safe_size(n):
-    """
-    Is the size of FFT such that FFTPACK can handle it in single precision
-    with sufficient accuracy?
-
-    Composite numbers of 2, 3, and 5 are accepted, as FFTPACK has those
-    """
-    n = int(n)
-
-    if n == 0:
-        return True
-
-    # Divide by 3 until you can't, then by 5 until you can't
-    for c in (3, 5):
-        while n % c == 0:
-            n //= c
-
-    # Return True if the remainder is a power of 2
-    return not n & (n-1)
-
-
-def _datacopied(arr, original):
-    """
-    Strict check for `arr` not sharing any data with `original`,
-    under the assumption that arr = asarray(original)
-
-    """
-    if arr is original:
-        return False
-    if not isinstance(original, numpy.ndarray) and hasattr(original, '__array__'):
-        return False
-    return arr.base is None
-
-
-def _asfarray(x):
-    """Like numpy asfarray, except that it does not modify x dtype if x is
-    already an array with a float dtype, and do not cast complex types to
-    real."""
-    if hasattr(x, "dtype") and x.dtype.char in numpy.typecodes["AllFloat"]:
-        # 'dtype' attribute does not ensure that the
-        # object is an ndarray (e.g. Series class
-        # from the pandas library)
-        if x.dtype == numpy.half:
-            # no half-precision routines, so convert to single precision
-            return numpy.asarray(x, dtype=numpy.float32)
-        return numpy.asarray(x, dtype=x.dtype)
-    else:
-        # We cannot use asfarray directly because it converts sequences of
-        # complex to sequence of real
-        ret = numpy.asarray(x)
-        if ret.dtype == numpy.half:
-            return numpy.asarray(ret, dtype=numpy.float32)
-        elif ret.dtype.char not in numpy.typecodes["AllFloat"]:
-            return numpy.asfarray(x)
-        return ret
-
-
-def _fix_shape(x, n, axis):
-    """ Internal auxiliary function for _raw_fft, _raw_fftnd."""
-    s = list(x.shape)
-    if s[axis] > n:
-        index = [slice(None)]*len(s)
-        index[axis] = slice(0,n)
-        x = x[tuple(index)]
-        return x, False
-    else:
-        index = [slice(None)]*len(s)
-        index[axis] = slice(0,s[axis])
-        s[axis] = n
-        z = zeros(s,x.dtype.char)
-        z[tuple(index)] = x
-        return z, True
-
-
 def fft(x, n=None, axis=-1, overwrite_x=False):
     """
     Return discrete Fourier transform of real or complex sequence.
